Basic test app for V4L2 M2M API
Main test has been for decode on a Pi with the bcm2835-codec driver
./m2m -d /dev/video0 -i foo.h264 -o decode.yuv
Input format adopts the default of H264.
Output format should be automatically detected and adopted.

It should also encode using
./m2m -d /dev/video1 -i source.yuv -o foo.h264
The source will currently need to be in the default format of
640x480 I420 (nothing sets the input format).
(Not been tested, but v4l2-ctl was encoding before decode was
worked on).