Seeking test.
Reworks to use the MPLANE variants.

The codec now expects framed data, therefore add hard coded offsets for
frame boundaries in the Big Buck Bunny clip also added.

It plays a number of input frames, and then does STREAMOFF and STREAMON
as if seeking, and starts again at a different point in the stream.

Test with:
./m2m -d /dev/video10 -i 720p.h264 -f -c 120 -o test.yuv